requirements.txt
numpy
nibabel
matplotlib
pytorch 
pandas
